% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreakinference.R
\name{outbreak.calc.cases}
\alias{outbreak.calc.cases}
\title{Calculate case count for an outbreak}
\usage{
outbreak.calc.cases(model, params, time)
}
\arguments{
\item{model}{model returned by \code{\link[mgcv]{gam}} or \code{\link[mgcv]{gamm}}, with a single parameter (time)}

\item{params}{model parameter matrix}

\item{time}{vector of time values at which the model will be evaluated}
}
\value{
time series of cumulative case counts
}
\description{
This is useful as \code{outcome} for outbreak.estimate.scalars.
}
