% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreakinference.R
\name{outbreak.estimate.scalars.confints}
\alias{outbreak.estimate.scalars.confints}
\title{Calculates confidence intervals for results of simulation performed by \code{\link{outbreak.estimate.scalars.sample}}}
\usage{
outbreak.estimate.scalars.confints(predictions, level = 0.95)
}
\arguments{
\item{predictions}{data frame of predictions as returned by \code{\link{outbreak.estimate.scalars.sample}}}

\item{level}{confidence level for calculated confidence intervals}
}
\value{
data.frame of confidence intervals, as returned by \code{\link{outbreak.estimate.scalars}}
}
\description{
Calculates confidence intervals for results of simulation performed by \code{\link{outbreak.estimate.scalars.sample}}
}
\keyword{internal}
