% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreakpredict.R
\name{outbreak.predict.scalars}
\alias{outbreak.predict.scalars}
\title{Calculates confidence intervals for scalar predicted from generalized additive (mixed) model of an outbreak}
\usage{
outbreak.predict.scalars(model, time, outcomes, nsim = 100, level = 0.95)
}
\arguments{
\item{model}{model returned by \code{\link[mgcv]{gam}} or \code{\link[mgcv]{gamm}}, with a single parameter (time)}

\item{time}{vector of time values at which the model will be evaluated}

\item{outcomes}{function returning calculated scalar outcomes, as described above}

\item{nsim}{number of simulations to run}

\item{level}{confidence level for predictions}
}
\value{
data frame of predictions, as described above
}
\description{
This function performs a series of Monte Carlo simulations of a GAM/GAMM outbreak model.
For each simulated outbreak, it calls \code{outcomess} to calculate scalar
outcomes for the simulated outbreak (for example, the time of outbreak peak).
It then calculates and returns the confidence interval of each simulated scalar
outcome across all simulations.
}
