% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreakpredict.R
\name{outbreak.predict.timeseries.confints}
\alias{outbreak.predict.timeseries.confints}
\title{Calculates confidence intervals for results of simulation performed by \code{\link{outbreak.predict.timeseries.sim}}}
\usage{
outbreak.predict.timeseries.confints(predictions, level = 0.95)
}
\arguments{
\item{predictions}{data frame of predictions as returned by \code{\link{outbreak.predict.timeseries.sim}}}

\item{level}{confidence level for calculated confidence intervals}
}
\value{
data.frame of confidence intervals, as returned by \code{\link{outbreak.predict.timeseries}}
}
\description{
Calculates confidence intervals for results of simulation performed by \code{\link{outbreak.predict.timeseries.sim}}
}
\keyword{internal}
