% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreakinference.R
\name{outbreak.predict.timeseries}
\alias{outbreak.predict.timeseries}
\title{Calculates confidence intervals for time series predicted from generalized additive (mixed) model of an outbreak}
\usage{
outbreak.predict.timeseries(model, time, outcome, nsim = 1000, level = 0.95)
}
\arguments{
\item{model}{model returned by \code{\link[mgcv]{gam}} or \code{\link[mgcv]{gamm}}, with a single parameter (time)}

\item{time}{vector of time values at which the model will be evaluated}

\item{outcome}{function returning calculated outcome time series, as described above}

\item{nsim}{number of simulations to run}

\item{level}{confidence level for returned predictions}
}
\value{
data frame of predictions, as described above
}
\description{
This function performs a series of Monte Carlo simulations of a GAM/GAMM outbreak model.
For each simulated outbreak, it calls \code{outcome} to calculate a time series for the
simulated outbreak (for example, the number of cumulative cases vs time).
It then calculates and returns the confidence interval of the simulated time series at
each time point across all simulations
}
