% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreakinference.R
\docType{package}
\name{outbreakinference-package}
\alias{outbreakinference-package}
\title{Stochastic modeling of infectious disease outbreak characteristics}
\description{
This package lets you made estimates (with confidence intervals) of characteristics of infectious
disease outbreaks.
}
\details{
It does this by modeling disease outbreaks using generalized additive (mixed) models (GAM/GAMM), then
using stochastic simulation based on the GAM/GAMM model to predict outbreak characteristics.

For example, this package can be used to model timing of outbreak onset, peak, or offset,
as well as outbreak cumulative incidence over time.

The package can model two types of outbreak characteristics: scalar characteristics, which are single-value
measures of an outbreak (for example, timing of peak) and time series characteristics, which are functions of
time (for example, cumulative incidence count over time)

For each outbreak characteristic, the package produces median and confidence interval estimates.

Typical use of this package begins by using the package \code{\link{mgcv}} to obtain
a GAM/GAMM model of the outbreak, followed by calling either \code{\link{outbreak.predict.scalars}} or
\code{\link{outbreak.predict.timeseries}} to obtain confidence intervals on the
desired scalar/timeseries outcomes of the outbreak

Both \code{\link{outbreak.predict.scalars}} and \code{\link{outbreak.predict.timeseries}}
allow simulation of arbitrary outbreak characteristics, by passing a function that calculates
the desired characteristics into \code{\link{outbreak.predict.scalars}} or \code{\link{outbreak.predict.timeseries}}.

For convenience, this package also includes \code{\link{outbreak.calc.thresholds}}, which can
be use in conjunction with \code{\link{outbreak.predict.scalars}} to predict timing of outbreak onset and
offset, as well as \code{\link{outbreak.calc.cum}}, which can be used in conjuction with
\code{\link{outbreak.predict.timeseries}} to predict cumulative incidence vs time for the outbreak.
}
\examples{
# Simulate an outbreak for analysis
cases = rpois(52, c(rep(1, 13), seq(1, 50, length.out=13), seq(50, 1, length.out=13), rep(1, 13)))
data = data.frame(cases=cases, time=seq(0, 51))

# Generate GAM model for outbreak; see mgcv for details
library(mgcv)
model = gam(cases ~ s(time, k=20, bs="cp", m=3), family=poisson, data=data)

# Generate time series at which model will be evaluated for predictions
# For the most part, you want the time span of the outbreak data divided into
# small increments (here, eps)
eps = .05
modelTime = seq(min(data$time) - 1 + eps, max(data$time), eps)

# Predict cumulative incidence count time series
cumCases = outbreak.predict.timeseries(
  model, modelTime,
  outbreak.calc.cum(timedelta=1), level=.95
)

# Predict time when outbreak crosses 5\% and 95\% of cumulative case count
thresholds = outbreak.predict.scalars(
  model, modelTime,
  outbreak.calc.thresholds(onset=0.05, offset=0.95), level=.95
)

# Plot cumulative incidence predictions and thresholds
library(ggplot2)
ggplot(cbind(cumCases, data.frame(time=modelTime))) +
  geom_ribbon(aes(x=time, ymin=lower, ymax=upper), fill=grey(.75)) +
  geom_line(aes(x=time, y=median)) +
  annotate("rect",
    xmin=thresholds$onset.lower,
    xmax=thresholds$onset.upper,
    ymin=-Inf, ymax=Inf, alpha=.25) +
  annotate("rect",
    xmin=thresholds$offset.lower,
    xmax=thresholds$offset.upper,
    ymin=-Inf, ymax=Inf, alpha=.25) +
  annotate("segment", x=-Inf, xend=Inf, y=0.05, yend=0.05) +
  annotate("segment", x=-Inf, xend=Inf, y=0.95, yend=0.95) +
 labs(x="Time", y="Relative cumulative incidence")

}
\author{
Ben Artin \email{ben@artins.org}
}
