% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{pspline.inference}
\alias{pspline.inference}
\alias{pspline.inference-package}
\title{Inference using penalized basis splines (P-splines) in a generalized additive model (GAM), with applications in
infectious disease outbreak modeling}
\description{
This package lets you make point and interval estimates of outcomes modeled with a non-linear P-spline GAM.
}
\details{
Applications in infectious disease outbreak modeling include estimating of outbreak onset, peak, or offset,
as well as outbreak cumulative incidence over time.

The package can model two types of outcomes: scalar outcomes, which are single-value
outcome measures (for example, timing of outbreak peak) and time series characteristics, which are functions of
time (for example, infection incidence over time)

For each outcome measure, the package produces median and confidence interval estimates.

Typical use of this package begins by using the package \code{\link{mgcv}} to obtain
a GAM/GAMM model of the process under investigation (such as an infectious disease outbreak), followed by
calling either \code{\link{pspline.estimate.scalars}} or \code{\link{pspline.estimate.timeseries}} to
obtain confidence intervals on the desired outcome measure

Both \code{\link{pspline.estimate.scalars}} and \code{\link{pspline.estimate.timeseries}}
allow computation of arbitrary outcome measures, by passing a function that calculates
the desired outcome measure into \code{\link{pspline.estimate.scalars}} or \code{\link{pspline.estimate.timeseries}}.

For convenience, this package also includes several utilities specifically aimed at modeling of
infectious disease outbreaks, such as \code{\link{pspline.outbreak.cases}} and \code{\link{pspline.outbreak.cumcases}}
(for estimation of incidence and cumulative incidence), and \code{\link{pspline.outbreak.thresholds}}, for estimation
of outbreak onset and offset.
}
\examples{
# Simulate an outbreak for analysis
cases = rpois(52, c(rep(1, 13), seq(1, 50, length.out=13), seq(50, 1, length.out=13), rep(1, 13)))
data = data.frame(cases=cases, time=seq(0, 51))

# Generate GAM model for outbreak; see mgcv for details
library(mgcv)
model = gam(cases ~ s(time, k=20, bs="cp", m=3), family=poisson, data=data)

# Generate time series at which model will be evaluated for estimates
# Usually you want this to be the same as the time interval that your observations are in, except
# divided into small increments (here, eps)
eps = .05
predictors = data.frame(time=seq(min(data$time) - 0.5, max(data$time) + 0.5 - eps, by=eps))

# Estimate cumulative incidence
cumCases = pspline.estimate.timeseries(
  model, predictors,
  pspline.outbreak.cumcases, level=.95
)

# Estimate time when outbreak crosses 5\% and 95\% of cumulative case count
onsetThreshold = 0.05
offsetThreshold = 1 - onsetThreshold
thresholds = pspline.estimate.scalars(
  model, predictors,
  pspline.outbreak.thresholds(onset=onsetThreshold, offset=offsetThreshold), level=.95
)

# Plot cumulative incidence estimates and threshold estimates
library(ggplot2)
ggplot(cumCases) +
  geom_ribbon(aes(x=time, ymin=cumcases.lower, ymax=cumcases.upper), fill=grey(.75)) +
  geom_line(aes(x=time, y=cumcases.median)) +
  annotate("rect",
    xmin=thresholds$onset.lower,
    xmax=thresholds$onset.upper,
    ymin=-Inf, ymax=Inf, alpha=.25) +
  annotate("rect",
    xmin=thresholds$offset.lower,
    xmax=thresholds$offset.upper,
    ymin=-Inf, ymax=Inf, alpha=.25) +
  annotate("segment", x=-Inf, xend=Inf, y=onsetThreshold, yend=onsetThreshold) +
  annotate("segment", x=-Inf, xend=Inf, y=offsetThreshold, yend=offsetThreshold) +
 labs(x="Time", y="Relative cumulative incidence")

}
\author{
Ben Artin \email{ben@artins.org}
}
