% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak.R
\name{pspline.outbreak.cases}
\alias{pspline.outbreak.cases}
\title{Calculate incidence for an outbreak}
\usage{
pspline.outbreak.cases(model, params, predictors)
}
\arguments{
\item{model}{model returned by \code{\link[mgcv]{gam}} or \code{\link[mgcv]{gamm}}}

\item{params}{model parameter matrix}

\item{predictors}{data frame of predictor values at which the model will be evaluated}
}
\value{
data frame of predictor values with corresponding incidence estimates in \code{$cases}
}
\description{
This is useful as \code{outcome} for \code{\link{pspline.estimate.scalars}}.
}
