% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak.R
\name{pspline.outbreak.cumcases}
\alias{pspline.outbreak.cumcases}
\title{Calculate cumulative case count for an outbreak}
\usage{
pspline.outbreak.cumcases(model, params, predictors)
}
\arguments{
\item{model}{model returned by \code{\link[mgcv]{gam}} or \code{\link[mgcv]{gamm}}, with a single parameter (time)}

\item{params}{model parameter matrix}

\item{time}{vector of time values at which the model will be evaluated}
}
\value{
time series of relative cumulative cases
}
\description{
This is useful as \code{outcome} for pspline.estimate.timeseries.
}
