% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreakpredict.R
\name{randomMVN}
\alias{randomMVN}
\title{Sample from a multivariate normal distribution}
\usage{
randomMVN(mu, sig, nsim)
}
\arguments{
\item{mu}{matrix of means}

\item{sig}{matrix of covariances}

\item{nsim}{number of samples to draw}
}
\value{
matrix of sampled values
}
\description{
Sample from a multivariate normal distribution
}
\keyword{internal}
