% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreakinference.R
\name{outbreak.calc.cases}
\alias{outbreak.calc.cases}
\title{Predict case count for an outbreak}
\usage{
outbreak.calc.cases(timedelta = 1)
}
\arguments{
\item{timedelta}{time step of the data described by \code{model}; note that this is the
time step of the original time series from which \code{model} was obtained, not the
(potentially different) time step at which model predictions are being evaluated}
}
\value{
time series of cumulative case counts
}
\description{
This is useful as \code{outcome} for outbreak.predict.scalars.
}
