% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreakinference.R
\name{outbreak.calc.thresholds}
\alias{outbreak.calc.thresholds}
\title{This is useful as \code{outcomes} for outbreak.estimate.scalars.}
\usage{
outbreak.calc.thresholds(onset = NA, offset = NA)
}
\arguments{
\item{onset}{onset threshold as fraction of total outbreak case count}

\item{offset}{offset threshold as fraction of total outbreak case count}
}
\value{
data frame with columns \code{onset} and \code{offset} representing time
when the outbreak crossed onsed and offset thresholds
}
\description{
This is useful as \code{outcomes} for outbreak.estimate.scalars.
}
