% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreakpredict.R
\name{outbreak.calc.thresholds}
\alias{outbreak.calc.thresholds}
\title{This is useful as \code{quant} for outbreak.predict.scalars.}
\usage{
outbreak.calc.thresholds(model, params, time, onset = 0.05, offset = 0.95)
}
\arguments{
\item{model}{GAM/GAMM model for the outbreak}

\item{params}{model parameters for the outbreak model}

\item{time}{vector of time values at which the model will be evaluated}

\item{onset}{onset threshold as fraction of total outbreak case count}

\item{offset}{offset threshold as fraction of total outbreak case count}
}
\value{
data frame with columns \code{onset} and \code{offset} representing time
when the outbreak crossed onsed and offset thresholds
}
\description{
This is useful as \code{quant} for outbreak.predict.scalars.
}
