% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreakinference.R
\name{outbreak.estimate.timeseries.confints}
\alias{outbreak.estimate.timeseries.confints}
\title{Calculates confidence intervals for results of simulation performed by \code{\link{outbreak.estimate.timeseries.sample}}}
\usage{
outbreak.estimate.timeseries.confints(samplesions, level = 0.95)
}
\arguments{
\item{level}{confidence level for calculated confidence intervals}

\item{samples}{data frame of samples as returned by \code{\link{outbreak.estimate.timeseries.sample}}}
}
\value{
data.frame of confidence intervals, as returned by \code{\link{outbreak.estimate.timeseries}}
}
\description{
Calculates confidence intervals for results of simulation performed by \code{\link{outbreak.estimate.timeseries.sample}}
}
\keyword{internal}
