% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak.R
\name{pspline.outbreak.cases}
\alias{pspline.outbreak.cases}
\title{Calculate case count for an outbreak}
\usage{
pspline.outbreak.cases(model, params, predictors)
}
\arguments{
\item{model}{model returned by \code{\link[mgcv]{gam}} or \code{\link[mgcv]{gamm}}}

\item{params}{model parameter matrix}

\item{time}{vector of time values at which the model will be evaluated}
}
\value{
time series of cumulative case counts
}
\description{
This is useful as \code{outcome} for pspline.estimate.scalars.
}
